+++ src/Azure/Storage.Net.Microsoft.Azure.DataLake.Store/Gen2/IAzureDataLakeGen2Storage.cs
@@ -1,7 +1,7 @@
 ﻿using Storage.Net.Blobs;
 using Storage.Net.Microsoft.Azure.DataLakeGen2.Store.Gen2.Interfaces;
 
-namespace Storage.Net.Microsoft.Azure.DataLakeGen2.Store
+namespace Storage.Net.Microsoft.Azure.DataLake.Store.Gen2
 {
    public interface IAzureDataLakeGen2Storage : IBlobStorage
    {
