+++ src/Azure/Storage.Net.Microsoft.Azure.DataLake.Store/Gen2/Interfaces/IHttpClientWrapper.cs
