+++ src/Azure/Storage.Net.Microsoft.Azure.DataLake.Store/Gen2/Models/AclItem.cs
@@ -1,4 +1,4 @@
-﻿namespace Storage.Net.Microsoft.Azure.DataLakeGen2.Store.Gen2.Models
+﻿namespace Storage.Net.Microsoft.Azure.DataLake.Store.Gen2.Models
 {
    public class AclItem
    {
