+++ src/Azure/Storage.Net.Microsoft.Azure.DataLake.Store/Gen2/Models/DataLakeGen2Exception.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
 
-namespace Storage.Net.Microsoft.Azure.DataLakeGen2.Store.Gen2.Models
+namespace Storage.Net.Microsoft.Azure.DataLake.Store.Gen2.Models
 {
    public class DataLakeGen2Exception : Exception
    {
