--- src/Azure/Storage.Net.Microsoft.Azure.DataLakeGen2.Store/Storage.Net.Microsoft.Azure.DataLakeGen2.Store.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-   <PropertyGroup>
-      <TargetFrameworks>netstandard1.6</TargetFrameworks>
-      <AssemblyOriginatorKeyFile>../../storage.snk</AssemblyOriginatorKeyFile>
-      <SignAssembly>true</SignAssembly>
-      <RepositoryUrl>https://github.com/aloneguid/storage</RepositoryUrl>
-      <RepositoryType>GitHub</RepositoryType>
-      <Version>1.0.0</Version>
-      <FileVersion>1.0.0.0</FileVersion>
-      <AssemblyVersion>1.0.0.0</AssemblyVersion>
-      <PackageLicenseUrl>https://github.com/aloneguid/storage/blob/master/LICENSE</PackageLicenseUrl>
-      <Copyright>Copyright (c) 2015-2017 by Ivan Gavryliuk</Copyright>
-      <PackageIconUrl>http://i.isolineltd.com/nuget/storage.png</PackageIconUrl>
-      <PackageProjectUrl>https://github.com/aloneguid/storage</PackageProjectUrl>
-      <Authors>Ivan Gavryliuk (@aloneguid)</Authors>
-      <Description>Provides Azure Data Lake Store integration for Storage.Net library</Description>
-      <LangVersion>latest</LangVersion>
-      <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-      <WarningsAsErrors />
-   </PropertyGroup>
-   <PropertyGroup Condition="'$(Configuration)' == 'Release'">
-      <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-   </PropertyGroup>
-   <ItemGroup>
-      <PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="3.19.8" />
-      <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-      <PackageReference Include="System.Net.Http" Version="4.3.4" />
-   </ItemGroup>
-   <ItemGroup>
-      <ProjectReference Include="..\..\Storage.Net\Storage.Net.csproj" />
-   </ItemGroup>
-</Project>
