+++ test/Storage.Net.Tests.Integration/Storage.Net.Tests.Integration.csproj
@@ -25,7 +25,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.DataLake.Store\Storage.Net.Microsoft.Azure.DataLake.Store.csproj" />
-    <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.DataLakeGen2.Store\Storage.Net.Microsoft.Azure.DataLakeGen2.Store.csproj" />
     <ProjectReference Include="..\..\src\General\Storage.Net.Ftp\Storage.Net.Ftp.csproj" />
     <ProjectReference Include="..\..\src\Storage.Net\Storage.Net.csproj" />
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.EventHub\Storage.Net.Microsoft.Azure.EventHub.csproj" />
