+++ test/Storage.Net.Tests/Storage.Net.Tests.csproj
@@ -13,7 +13,7 @@
       <WarningsAsErrors />
    </PropertyGroup>
    <ItemGroup>
-      <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.DataLakeGen2.Store\Storage.Net.Microsoft.Azure.DataLakeGen2.Store.csproj" />
+      <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.DataLake.Store\Storage.Net.Microsoft.Azure.DataLake.Store.csproj" />
       <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.Storage\Storage.Net.Microsoft.Azure.Storage.csproj" />
       <ProjectReference Include="..\..\src\Storage.Net\Storage.Net.csproj" />
    </ItemGroup>
